Added enum for toggle between how we offload cache after the forward pass
By default we count the layerwise executions in the forward pass
but in the case of using cuda graphs with trtllm, forward passes
two and greater don`t trigger our kvbm counts. Thus, we hook into
wait_for_save() which is called at the end of every forward pass,
bypassing the need to count the layers from the stream events.

wait_for_save() might be slower since when getting the stream of events
we only have to sync on the cuda events to know it is safe to
transfer. This can be optimized later.

test to prove the previous functionality holds

Added a test to prove trtllm worker + kvbm - cuda graphs
still work with the required changes.

Revert "Added enum for toggle between how we offload cache after the forward pass"

This reverts commit 0921d6b96519103fae5cc02bbdf1623886843286.

Added enums for the different offloading modes and added them to the kvbm configuration struct

fixing compile bugs

initial pass at api interface for passing a cuda event to trtllm

Signed-off-by: Kyle McGill <kmcgill@nvidia.com>

using new TRTLLM ForwardPassCallback api

clean up for pr

Revert "Added enums for the different offloading modes and added them to the kvbm configuration struct"

This reverts commit 3feb5cdb0b5cae326ab27927fa9d2f446309ca7c.

clean up for pr

whitespace

synchronizing on the cuda event instead of on the stream

Refactored the API to provide a callable to TRTLLM

Instead of returning a class which needs to have a
function implemented, the API now provides a callable
which will be called at the end of the forward pass.

Added feature to make DynamoKVBMConnectorWorker backward compatible

code rabbit suggestions

test file PR typo and formatting suggestions

adding test for with/without cuda graphs

fixing tests